requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
openai
beautifulsoup4
lxml
python-json-logger
